Moves blur function to UIView, as it is not only for UIImageView
