+++ Sources/SwifterSwift/UIKit/UIImageViewExtensions.swift
@@ -35,28 +35,6 @@ public extension UIImageView {
             }
         }.resume()
     }
-
-    /// SwifterSwift: Make image view blurry.
-    ///
-    /// - Parameter style: UIBlurEffectStyle (default is .light).
-    func blur(withStyle style: UIBlurEffect.Style = .light) {
-        let blurEffect = UIBlurEffect(style: style)
-        let blurEffectView = UIVisualEffectView(effect: blurEffect)
-        blurEffectView.frame = bounds
-        blurEffectView.autoresizingMask = [.flexibleWidth, .flexibleHeight] // for supporting device rotation
-        addSubview(blurEffectView)
-        clipsToBounds = true
-    }
-
-    /// SwifterSwift: Blurred version of an image view.
-    ///
-    /// - Parameter style: UIBlurEffectStyle (default is .light).
-    /// - Returns: blurred version of self.
-    func blurred(withStyle style: UIBlurEffect.Style = .light) -> UIImageView {
-        let imgView = self
-        imgView.blur(withStyle: style)
-        return imgView
-    }
 }
 
 #endif
