+++ Tests/UIKitTests/UIImageViewExtensionsTests.swift
@@ -36,24 +36,6 @@ final class UIImageViewExtensionsTests: XCTestCase {
         XCTAssertNil(failImageView.image)
         waitForExpectations(timeout: 15)
     }
-
-    func testBlur() {
-        let imageView = UIImageView(frame: CGRect(x: 0, y: 0, width: 50, height: 100))
-        imageView.blur(withStyle: .dark)
-
-        let blurView = imageView.subviews.first as? UIVisualEffectView
-        XCTAssertNotNil(blurView)
-        XCTAssertNotNil(blurView?.effect)
-        XCTAssertEqual(blurView?.frame, imageView.bounds)
-        XCTAssertEqual(blurView?.autoresizingMask, [.flexibleWidth, .flexibleHeight])
-        XCTAssert(imageView.clipsToBounds)
-    }
-
-    func testBlurred() {
-        let imageView = UIImageView()
-        let blurredImageView = imageView.blurred(withStyle: .extraLight)
-        XCTAssertEqual(blurredImageView, imageView)
-    }
 }
 
 #endif
